iface: move igmp code to a separate mod.
